pyproject.toml
[project]
name = "burmese-ocr"
version = "0.1.0"
description = "A tool for OCR of Burmese text"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pytesseract",
    "pillow",
    "streamlit",
    "opencv-python",
    "numpy",
]
